+++ Globals.lua
@@ -319,7 +319,7 @@ function addon.HookAchievementFrameOnShow()
     hooksecurefunc(funcName, addon.GUI.ShowHideTabs);
 end
 
-local function MakeMovable(frame, target)
+local function MakeMovable(frame, rememberLastPositionOption, target)
     if frame:IsMovable() then -- Do not hook it multiple times if another addon already made it movable
         return;
     end
@@ -335,22 +335,55 @@ local function MakeMovable(frame, target)
     end);
     frame:SetScript("OnMouseUp", function(frame, button)
         target:StopMovingOrSizing();
+        if addon.Options.db.Window.RememberLastPosition[rememberLastPositionOption] then
+            SavedData.RememberLastPosition = SavedData.RememberLastPosition or {};
+            SavedData.RememberLastPosition[rememberLastPositionOption] = {
+                X = target:GetLeft(),
+                Y = target:GetTop() - UIParent:GetTop()
+            };
+        end
     end);
 end
 
 function addon.MakeWindowMovable()
+    SavedData.RememberLastPosition = SavedData.RememberLastPosition or {};
+    if not SavedData.RememberLastPosition["AchievementWindow"] then
+        SavedData.RememberLastPosition["AchievementWindow"] = {
+            X = 96,
+            Y = -116
+        };
+    end
+    if not SavedData.RememberLastPosition["Calendar"] then
+        SavedData.RememberLastPosition["Calendar"] = {
+            X = 150,
+            Y = -80
+        };
+    end
+    if not SavedData.RememberLastPosition["DataManager"] then
+        SavedData.RememberLastPosition["DataManager"] = {
+            X = 150,
+            Y = -80
+        };
+    end
+
     if not addon.Options.db.Window.Movable then
         return;
     end
     if AchievementFrame and AchievementFrameHeader then
-        MakeMovable(AchievementFrame);
-        MakeMovable(AchievementFrameHeader, AchievementFrame);
+        local pos = SavedData.RememberLastPosition["AchievementWindow"];
+        AchievementFrame:SetPoint("TOPLEFT", pos.X, pos.Y);
+        MakeMovable(AchievementFrame, "AchievementWindow");
+        MakeMovable(AchievementFrameHeader, "AchievementWindow", AchievementFrame);
     end
     if addon.GUI.Calendar.Frame then
-        MakeMovable(addon.GUI.Calendar.Frame);
+        local pos = SavedData.RememberLastPosition["Calendar"];
+        addon.GUI.Calendar.Frame:SetPoint("TOPLEFT", pos.X, pos.Y);
+        MakeMovable(addon.GUI.Calendar.Frame, "Calendar");
     end
     if addon.GUI.DataManagerFrame then
-        MakeMovable(addon.GUI.DataManagerFrame);
+        local pos = SavedData.RememberLastPosition["DataManager"];
+        addon.GUI.DataManagerFrame:SetPoint("TOPLEFT", pos.X, pos.Y);
+        MakeMovable(addon.GUI.DataManagerFrame, "DataManager");
     end
 end