+++ Localization/frFR.ExtraWrathStrings.lua
@@ -1,0 +1,17 @@
+-- [[ Exported at 2022-09-01 19-21-14 ]] --
+-- [[ This code is automatically generated as an export from ]] --
+-- [[ an SQLite database and is not meant for manual edit. ]] --
+
+-- [[ Namespaces ]] --
+local _, addon = ...;
+addon.ExtraWrathStrings.frFR = {};
+local extraWrathStrings = addon.ExtraWrathStrings.frFR;
+
+function extraWrathStrings.Load(L)
+    if addon.IsNotWotLKClassic() then
+        return;
+    end
+
+
+end
+