+++ Localization/zhCN.ExtraWrathStrings.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2022-08-19 07-49-05 ]] --
+-- [[ Exported at 2022-09-01 19-21-14 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --