+++ CHANGELOG.md
@@ -6,10 +6,12 @@ All notable changes to this project will be documented in this file.
 - Compatibility with Zygor Guides Viewer for the Blizzard Achievements tab
 - Focused, Current Zone, Selected Zone and Excluded category to every tab (option to hide them in unwanted tabs to be added)
 - Refresh to World and Widget Event Reminder options
+- Options to keep the Achievement Window, Achievement Calendar and/or Data Manager at the last position between sessions
 
 ### Fixed
 - When an achievement is selected and a criteria is completed, the criteria should show as completed without having to toggle the achievement
 - World events should always show time data again (untill Blizzard breaks the API again)
+- Opening the Achievement Window during combar is now always possible, even with rebound micro button
 
 ## 39.3 - 2022-08-21
 ### Fixed